+++ CHANGELOG.md
@@ -1,5 +1,12 @@
 # CHANGELOG
 
+### Pre-release v4.0.0b7 - Release date: 2024-05-14
+
+* Reduced writing time in status updater background task
+* Improved documentation
+* Removed `listexp` view
+* Minor fixes
+
 ### Pre-release v4.0.0b6 - Release date: 2024-04-19
 
 * Added filter by autosubmit version in `/v4/experiments`